ai-tool/mcp_sql/pyproject.toml
[project]
name = "mysql-mcp-server"
version = "1.0.0"
description = "MySQL MCP Server for AI assistants"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "mysql-connector-python>=8.0.33",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
